fix support for numpy.float64 values in parameters arg (like sigma2=np.float64(123))

+++ bindings/Python/tests/pylibkriging_demo.py
@@ -7,7 +7,7 @@ def test_kriging():
     y = [f(xi) for xi in X]
 
     import pylibkriging as lk
-    k_py = lk.Kriging(y, X, "gauss", parameters={'sigma2': 1, 'is_theta_estim': False})
+    k_py = lk.Kriging(y, X, "gauss", parameters={'sigma2': np.float64(1), 'is_theta_estim': False})
     print(k_py.summary())
 
     x = np.arange(0, 1, 1 / 99)
